aws_cdk_serverless_stack/requirements.txt
-e .
aws.cdk.aws_lambda
aws.cdk.aws_dynamodb
aws_cdk.aws_iam